Upload createNoisyDataMFCC.py which creates a data.csv with MFCC values of the samplesongs plus it creates noisy songs and use it immediatly as MFCC values and delete it afterwards.
Upload createModelFunction.py which use gridSearcher to find the best hypervalues for the model. The model itself is has change parameter to run faster. This programm tries every given value to make model and those values with the best output is shown in the console. This take a lot of time.