Upload equalizeGenreSongs.py which will drop rows from a genre if there to many values of it.
Upload makeCSV.py which is a function that combine the Song Splitter and the Gaussian Maker in one. It creates a CSV file which will include the MFCC values. It creates parts of standard 5 seconds and gaussian noise in one. Without double saving them up, instead it saves it just temporarly